requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --no-emit-project
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via google-genai
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
attrs==25.4.0
    # via aiohttp
aws-lambda-powertools==3.24.0
    # via emailbot
boto3==1.42.54
    # via
    #   emailbot
    #   moto
botocore==1.42.54
    # via
    #   boto3
    #   moto
    #   s3transfer
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
coverage==7.13.4
    # via pytest-cov
cryptography==46.0.5
    # via
    #   google-auth
    #   moto
distro==1.9.0
    # via google-genai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.48.0
    # via google-genai
google-genai==1.64.0
    # via emailbot
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via moto
jmespath==1.1.0
    # via
    #   aws-lambda-powertools
    #   boto3
    #   botocore
markdown==3.10.2
    # via emailbot
markupsafe==3.0.3
    # via
    #   jinja2
    #   werkzeug
moto==5.1.21
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   emailbot
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   pytest-cov
    #   pytest-mock
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto
pyyaml==6.0.3
    # via responses
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
    #   moto
    #   responses
responses==0.26.0
    # via moto
rsa==4.9.1
    # via google-auth
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via google-genai
tenacity==9.1.4
    # via google-genai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   aws-lambda-powertools
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   botocore
    #   requests
    #   responses
websockets==15.0.1
    # via google-genai
werkzeug==3.1.6
    # via moto
xmltodict==1.0.3
    # via moto
yarl==1.22.0
    # via aiohttp
